Se añadieron las clases, las clases estan en diferentes cabeceras, hay una cabecera que tiene las clases y otra las funciones, esto para evitar que el main este saturado, se implemento la funcion de buscar un enemigo (esta en proceso)
