+++ clasetienda.h
@@ -1,0 +1,22 @@
+#ifndef CLASETIENDA_H
+#define CLASETIENDA_H
+#include <iostream>
+using namespace std;
+
+//CREACION DE LA CLASE TIENDA
+class Tienda {
+	//ATRIBUTOS
+	private:
+		Objeto objetosEnTienda[10];
+	
+	//FUNCIONES
+	public:
+		//CONSTRUCTOR
+		Tienda() {
+			//objetosEnTienda[10] = 
+		}
+		
+		virtual void mostrarObjetos();
+		virtual void venderObjetosAlHeroe();
+};
+#endif
