+++ funciones.h
@@ -1,0 +1,7 @@
+#ifndef FUNCIONES_H
+#define FUNCIONES_H
+
+//SE AGREGAN TODAS LAS FUNCIONES PARA TENERLAS EN UNA SOLA CABECERA
+#include "funcionesbosque.h"
+
+#endif
