+++ funcionesbosque.h
@@ -1,0 +1,7 @@
+#ifndef FUNCIONESBOSQUE_H
+#define FUNCIONESBOSQUE_H
+#include <iostream>
+using namespace std;
+
+void caminar(Heroe jugador);
+#endif
